# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CamWatch — Multi-tenant IP camera monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new tenant + owner account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── DASHBOARD ──────────────────────────────────────────────────────────
...
  /cameras:
    get:
      operationId: listCameras
      tags: [cameras]
      summary: List all cameras for the tenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CameraList"
    post:
      operationId: createCamera
      tags: [cameras]
      summary: Register a new camera
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCameraBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"
  /cameras/{id}:
    get:
      operationId: getCamera
      tags: [cameras]
      summary: Get camera details
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"
    patch:
      operationId: updateCamera
      tags: [cameras]
      summary: Update camera
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCameraBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"
    delete:
      operationId: deleteCamera
      tags: [cameras]
      summary: Delete camera
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
  /cameras/{id}/status:
    patch:
      operationId: updateCameraStatus
      tags: [cameras]
      summary: Update camera online/offline status
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCameraStatusBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"

  # ─── EVENTS ─────────────────────────────────────────────────────────────
...
  /edge-agents:
    get:
      operationId: listEdgeAgents
      tags: [edge-agents]
      summary: List edge agents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdgeAgentList"
    post:
      operationId: createEdgeAgent
      tags: [edge-agents]
      summary: Register an edge agent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEdgeAgentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdgeAgent"
  /edge-agents/{id}:
    patch:
      operationId: updateEdgeAgent
      tags: [edge-agents]
      summary: Update edge agent
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEdgeAgentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdgeAgent"
    delete:
      operationId: deleteEdgeAgent
      tags: [edge-agents]
      summary: Delete edge agent
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
  /edge-agents/{id}/heartbeat:
    post:
      operationId: edgeAgentHeartbeat
      tags: [edge-agents]
      summary: Edge agent heartbeat ping
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdgeAgent"

  # ─── USERS ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    RegisterBody:
      type: object
      properties:
        tenantName:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [tenantName, name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    RefreshBody:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [accessToken, refreshToken, user]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        tenantId:
          type: integer
        tenantName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, tenantId, tenantName, createdAt]

    # ─── DASHBOARD ──────────────────────────────────────────────────────
...
    Camera:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        rtspUrl:
          type: string
        ip:
          type: string
        username:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        onvifEnabled:
          type: boolean
        mode:
          type: string
          enum: [cloud_events, edge_ai]
        status:
          type: string
          enum: [online, offline, unknown]
        location:
          type: string
        notes:
          type: string
        lastSeenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, name, mode, status, onvifEnabled, createdAt]
    CameraList:
      type: array
      items:
        $ref: "#/components/schemas/Camera"
    CreateCameraBody:
      type: object
      properties:
        name:
          type: string
        rtspUrl:
          type: string
        ip:
          type: string
        username:
          type: string
        password:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        onvifEnabled:
          type: boolean
        mode:
          type: string
          enum: [cloud_events, edge_ai]
        location:
          type: string
        notes:
          type: string
      required: [name, mode]
    UpdateCameraBody:
      type: object
      properties:
        name:
          type: string
        rtspUrl:
          type: string
        ip:
          type: string
        username:
          type: string
        password:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        onvifEnabled:
          type: boolean
        mode:
          type: string
          enum: [cloud_events, edge_ai]
        location:
          type: string
        notes:
          type: string
    UpdateCameraStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [online, offline, unknown]
      required: [status]

    # ─── EVENT ──────────────────────────────────────────────────────────
...
    EdgeAgent:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        token:
          type: string
        platform:
          type: string
          enum: [linux, windows, docker, raspberry_pi]
        version:
          type: string
        status:
          type: string
          enum: [online, offline, unknown]
        lastHeartbeatAt:
          type: string
          format: date-time
        cameraIds:
          type: array
          items:
            type: integer
        config:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, name, token, status, createdAt]
    EdgeAgentList:
      type: array
      items:
        $ref: "#/components/schemas/EdgeAgent"
    CreateEdgeAgentBody:
      type: object
      properties:
        name:
          type: string
        platform:
          type: string
          enum: [linux, windows, docker, raspberry_pi]
        cameraIds:
          type: array
          items:
            type: integer
        config:
          type: object
          additionalProperties: true
      required: [name]
    UpdateEdgeAgentBody:
      type: object
      properties:
        name:
          type: string
        platform:
          type: string
          enum: [linux, windows, docker, raspberry_pi]
        version:
          type: string
        cameraIds:
          type: array
          items:
            type: integer
        config:
          type: object
          additionalProperties: true

    # ─── USERS ──────────────────────────────────────────────────────────
...