# lib/api-zod/src/generated/types/updateEdgeAgentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CamWatch — Multi-tenant IP camera monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEdgeAgentBodyConfig } from "./updateEdgeAgentBodyConfig";
import type { UpdateEdgeAgentBodyPlatform } from "./updateEdgeAgentBodyPlatform";

export interface UpdateEdgeAgentBody {
  name?: string;
  platform?: UpdateEdgeAgentBodyPlatform;
  version?: string;
  cameraIds?: number[];
  config?: UpdateEdgeAgentBodyConfig;
}